[2/7] WTBS S5.2 - Remove a test unit (msi:suminfo).
Removing a test unit would not impact the other tests... except that it
requires modifying the makefile which could impact the other tests.

So the TestBot should rerun all the module's test units, including the
one added by the previous part, and excluding the one that was removed.

----- TestWTBS -----
p tests.TestUnits msi:dummy msi
; The msi tests are not reliable so ignore test failures.

a tests.report.GrepV msi:suminfo
----- TestWTBS -----