[1/2] WTBS S7.1 - Pseudo TestBot patch.
Make sure the TestBot ignores non-Wine patches rather than complain
that they fail to apply.
Here it should detect that this patch is for a Wine-related but non-Wine
repository and thus ignore this patch rather than treat it as a Wine
patch adding a new file.

----- TestWTBS -----
p patch.Disposition Part is not a Wine patch
----- TestWTBS -----