[2/2] WTBS BUG:48034 S7.2 - Second pseudo TestBot patch.
FIXME: The TestBot does not check patches against non-Wine repositories.
       Note that this does not impact S7.1 because nothing in the patch
       proves conclusively that it is a Wine patch. But the TestBot then
       adds the new file to the list of Wine files and thus mistakenly
       considers S7.2 to be a Wine patch.

The TestBot should identify this patch as a non-Wine patch.
When a previous part of a non-Wine patchset adds a file it's important
that later parts don't think that file is part of Wine as this would
cause them to think this is a Wine patch.

----- TestWTBS -----
p patch.Disposition Part is not a Wine patch
----- TestWTBS -----