[2/4] WTBS BUG:48034 S6.2 - Patch to a non-Wine file.
FIXME: The TestBot adds the S6.1 new file to the list of Wine files
       and thus mistakenly considers S6.2 to be a Wine patch.

Since S6.1 is not considered to be a Wine patch, modifying the file it
creates should not count as modifying a Wine file. So the TestBot should
also treat this as a non-Wine patch.

----- TestWTBS -----
p patch.Disposition Part is not a Wine patch
----- TestWTBS -----