[3/4] WTBS S6.3 - Wine patch which does not touch existing files.
The presence of a new Makefile.in file in the patch will trigger a
make_makefiles run which will add it to configure.ac, ensuring the new
directory is taken into account in the Wine build. Thus this qualifies
this patch as a Wine patch.
So the TestBot should verify that Wine still compiles.

----- TestWTBS -----
p build.HasTask 0
p win.HasTask 0
a wine.log.GrepV Task: tests
----- TestWTBS -----