WTBS Simple non-test patch to a dll with no tests (strmbase+quartz:acmwrapper).
This checks that we don't try to rebuild the Windows test executable
just because the module has been patched.
This is particularly important as it would cause a failure in case the
module has no tests such as here.
Note that the patch to the quartz test is needed otherwise the patch
would be ignored by the Windows side altogether.

----- TestWTBS -----
p tests.TestUnits quartz:acmwrapper
p tests.TestFailures 0

a tests.report.GrepV ^strmbase
a win.log.GrepV strmbase_test
a wine.log.GrepV winetest\.exe .* strmbase:
----- TestWTBS -----