WTBS BUG:48031 Test handling of runaway build traces (advapi32:crypt).
FIXME The TestBot does not support limiting the task log size yet.
FIXME Once implemented modify the TestWTBS section to look for the
      "truncated task log" error.

The TestBot should limit the size of the build logs it retrieves in
case a build got into an infinite loop printing messages.
Ideally it would also interrupt the build as soon as the log exceeds
the maximum size, hence the slow write.

----- TestWTBS -----
p tests.TestUnits advapi32:crypt

a build.log.Grep 1000  123456789 123456789
a wine.log.Grep 1000  123456789 123456789
----- TestWTBS -----