WTBS BUG:48795 Add an mbox for testing HTML attachments.
FIXME: WinePatchesMLSubmit.pl handles this correctly but
       WinePatchesWebSubmit.pl mishandles attachments when producing the
       webpatches/ files that are the input to the TestBot.

The mbox contains an email where someone replied to a patch with a MUA
that produces HTML email.
In the plain text part the patch got quoted as expected and is thus
ignored. But the HTML part looks like it contains a patch but it should
still be ignored on the basis of the part's MIME type.

In itself the TestBot should probably ignore this patch as not having
an impact on Wine.

----- TestWTBS -----
p patch.Disposition Not a Wine patch
----- TestWTBS -----