WTBS Add a dll containing dots in its name (lib.and.dot:base).
Wine forces dlls that contain dots in their names to be implemented in
a folder ending in '.dll'. But the module name and test executables
don't have this extra 'extension' and thus deriving them from the
patched filename requires special treatment.

Check that the TestBot succeeds in running the test unit and that the
log analysis does not choke on the dots in the module name.

----- TestWTBS -----
p tests.TestUnits lib.and.dot:base
p tests.TestFailures 0

a tests.report.GrepV WTBS Uh\? Somebody called\?
a tests.report.Grep WTBS Hello from the lib\.and\.dot test
----- TestWTBS -----